requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
matplotlib
plotly
faiss-cpu
sentence-transformers
transformers
torch
pypdf
Pillow
python-dotenv
